Rewrite syscall timestamp
