+++ ckb-debugger/src/arch_unix.rs
@@ -1,3 +1,7 @@
 pub fn println(s: &str) {
     println!("{}", s);
 }
+
+pub fn timestamp() -> u64 {
+    std::time::SystemTime::now().duration_since(std::time::SystemTime::UNIX_EPOCH).unwrap().as_secs()
+}
