+++ ckb-debugger/src/syscall_timestamp.rs
@@ -1,0 +1,30 @@
+use ckb_vm::{
+    Error, Register, SupportMachine, Syscalls,
+    registers::{A0, A7},
+};
+use std::time::SystemTime;
+
+pub struct Timestamp {}
+
+impl Timestamp {
+    pub fn new() -> Self {
+        Self {}
+    }
+}
+
+impl<Mac: SupportMachine> Syscalls<Mac> for Timestamp {
+    fn initialize(&mut self, _machine: &mut Mac) -> Result<(), Error> {
+        Ok(())
+    }
+
+    fn ecall(&mut self, machine: &mut Mac) -> Result<bool, Error> {
+        let id = machine.registers()[A7].to_u64();
+        if id != 9001 {
+            return Ok(false);
+        }
+        let duration = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap();
+        let now = duration.as_nanos();
+        machine.set_register(A0, Mac::REG::from_u64(now as u64));
+        return Ok(true);
+    }
+}
